Fixed data loss modifying compressed disk images
Only the first changes to gzipped disk images were correctly written to disk
when the drive motor switched off. Any further changes were held in memory but
never flushed to the disk image. The in-memory changes were lost when the disk
image was ejected.

Special thanks to Geoff for helping me reproduce this elusive bug!